Clase Principal y Pokémon
Principal: creación de submenús.
Pokémon: Constructor y getNombre

+++ src/main/java/simulador/pokemon/Pokemon.java
@@ -2,8 +2,9 @@ package simulador.pokemon;
 
 public abstract class Pokemon {
     private String nombre;
-    private double salud;
+    private int salud;
     private int puntosDeAtaque;
     private String tipo;
     private String estado;
 }
+
